services/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ambara-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "ny-feoko-shared",
    "pipeline",
    "clip-extraction",
    "db-sync",
    "asr-training",
    "yt-download",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-dotenv>=1",
    "supabase>=2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/api"]
